テストが通るように修正プラス在庫受注の処理実装

+++ src/main/java/jp/co/mgsystems/yuricollection/gootscatalog/beans/Product.java
@@ -21,7 +21,7 @@ public class Product {
     // 価格
     private Integer price;
     // 在庫数
-    private Long stocks;
+    private Integer stocks;
     // 受注数
     private Long orders;
     // コメント
